@@ -1,18 +1,17 @@
 package apicross.springmvc.params;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @WebMvcTest(controllers = TestController.class)
 public class ParamNameDataBinderTests {
     @Autowired
@@ -35,7 +34,6 @@ public class ParamNameDataBinderTests {
     @Test
     public void arrayParameterBindingWorks() throws Exception {
         this.mockMvc.perform(MockMvcRequestBuilders.get("/test?param-2=123,456"))
-                .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(content().json("{\"param2\":[123,456]}"));
     }
